Implement js entry point with views, add two routes
We set up our js entry point which mostly just waits for domReady and
then reads our embedded data block whose `viewName` prop defines the
current view name as chosen by server routing.

We use Webpack's require context functionality to create a "views"
folder from which all js modules are loaded automatically into the
webpack javascript bundle.

Each module in the views folder exports a React component as default
export which we render as our view. We provide it with the former Spring
Controller model as component props.

Here our initial data is not used for anything but being stringified
back to JSON to show its there.

In pratice this can be all kinds of prepared data for the view
components.

Of course the setup also doesn't need to be view component centric. The
JSON data from the server can also be used as e.g. initial redux state.

Webpack
-------

We define a webpack build that reads src/main/js/main.js as entry point
which looks up the component from the view name it receives from
the server and renders that component into the #root element we defined
in the base template.