Implement AJAX support
Enhance our react view so that it can detect a special header we send
with the "whatwg-fetch" / native fetch.

If the view is invoked with that header set, it does not answer with the
base HTML document but instead just with the normally embedded data as
JSON reponse.

We can render the JSON response as new JSON data with our system.

We add a very simple pushState based history so that our AJAX/fetch
requests still lead to user navigatable URLs.